prevent immediate debug mode monaco load to colorize sql
